Create paralindrome.py
For tasks 1 thru 3 use programming language of your choice from: Java, C++, or Python. Do not use any libraries, only native functions.

(1) Provided as input the following: a) An array of characters of any length.  b) The length of the array.        The output is to determine if the array contains a palindrome (i.e. words or phrases that read forward and backwards the same way).

(2)Provided as input the following: a) An array of characters of any length.  b) The length of the array. The Output are the number of each vowel is in the input, and the relative positions of the vowels within the word. Example "parabola" - vowel "a" shows up 3 times and in positions 2, 4, and 8. vowel "o" shows up 1 time, and in position 6.
       
3) Write a function that can parse a text file, and then determine how many words are palindromes and the vowel count/position for every word on the text in the file. You can use the libraries related to opening/reading/closing files, and the previous written functions. If they need to be changed you can do that as well for this task.



4) Find the mistake(s) in the Java code bellow, and how can it be corrected? A Prime number is a number that can only be divided by 1 and itself, example 11,37. The operator %, is to get the remainder. Example 10 % 3 will result in 1. 4%2 will yield 0.

int IsItPrime(int number) {

   int i,j;
   int aux,count;
   
   aux = number;
   
   for (i = 0; i < number; i++) {
         aux = number % j;
         if (aux == 0)
                count++;        
   }
   if (count == 2)
     System.out.println("Prime");
   else
     System.out.println("not Prime");
 
}